[VECTOR_FLOAT16] Add support for vector float16 data type (#2899)
* Added support for vector float16 and test scenarios

* Refactored tests and added proper comments

* Update VectorFloat16Test.java

* Fixed getTypedefinition for null vector data

* Addressed comments

* Addressed comments

* Addressed comments

* Addressed comments

* Added test scenarios

* Fixed typedefinition handling
