+++ src/main/java/com/microsoft/sqlserver/jdbc/SQLServerCallableStatement.java
@@ -390,9 +390,20 @@ public class SQLServerCallableStatement extends SQLServerPreparedStatement imple
         checkClosed();
 
         registerOutParameter(index, sqlType);
-        inOutParam[index - 1].setOutScale(scale);
+
         if (microsoft.sql.Types.VECTOR == sqlType) {
-            inOutParam[index - 1].setValueLength(VectorUtils.getDefaultPrecision()); // default 32-bit (single precision) float
+            // For VECTOR type, the 3-arg registerOutParameter(index, sqlType, scale) overload
+            // repurposes the 'scale' parameter as the vector dimension count (e.g. 3 for a
+            // 3-dimensional vector). This is stored as valueLength so that the type definition
+            // is built as VECTOR(<dimensionCount>).
+            // The outScale is set to the default bytes-per-dimension (4, i.e. FLOAT32) since
+            // this overload does not accept an explicit bytes-per-dimension value. For FLOAT16
+            // vectors, use the 4-arg overload registerOutParameter(index, sqlType, precision, scale)
+            // where precision = dimension count and scale = bytes-per-dimension (2).
+            inOutParam[index - 1].setValueLength(scale);
+            inOutParam[index - 1].setOutScale(VectorUtils.getDefaultPrecision());
+        } else {
+            inOutParam[index - 1].setOutScale(scale);
         }
 
         loggerExternal.exiting(getClassNameLogging(), "registerOutParameter");
@@ -822,7 +833,8 @@ public class SQLServerCallableStatement extends SQLServerPreparedStatement imple
             if (returnValue == null) {
                 TypeInfo typeInfo = getterGetParam(index).getTypeInfo();
                 int precision = typeInfo.getPrecision();
-                VectorDimensionType scale = VectorUtils.getVectorDimensionType(typeInfo.getScale());
+                VectorDimensionType scale = VectorUtils.getVectorDimensionType(
+                        VectorUtils.getScaleByte(typeInfo.getScale()));
                 vector = new Vector(precision, scale, null);
             } else {
                 vector = VectorUtils.fromBytes((byte[]) returnValue);
