+++ src/main/java/com/microsoft/sqlserver/jdbc/VectorUtils.java
@@ -17,10 +17,10 @@ import microsoft.sql.Vector.VectorDimensionType;
 class VectorUtils {
 
     private static final int VECTOR_HEADER_LENGTH = 8;
-    private static final int BYTES_PER_FLOAT = 4;
+    private static final int BYTES_PER_FLOAT32 = 4;
     private static final byte SCALE_BYTE_FLOAT32 = 0x00;
-    // private static final int BYTES_PER_SHORT = 2;
-    // private static final byte SCALE_BYTE_FLOAT16 = 0x01;
+    private static final int BYTES_PER_FLOAT16 = 2;
+    private static final byte SCALE_BYTE_FLOAT16 = 0x01;
 
     /**
      * Converts a byte array to a Vector object. The byte array must contain the following:
@@ -44,16 +44,7 @@ class VectorUtils {
         }
 
         int objectCount = (bytes.length - getHeaderLength()) / bytesPerDimension; // 8 bytes for header
-        Object[] objectArray;
-        switch (vectorType) {
-            // case FLOAT16:
-            // objectArray = new Short[objectCount];
-            // break;
-            case FLOAT32:
-            default:
-                objectArray = new Float[objectCount];
-                break;
-        }
+        Object[] objectArray = new Float[objectCount];
 
         ByteBuffer buffer = ByteBuffer.wrap(bytes).order(ByteOrder.LITTLE_ENDIAN);
 
@@ -65,7 +56,12 @@ class VectorUtils {
         ((Buffer) buffer).position(getHeaderLength()); // Skip the first 8 bytes (header)
 
         for (int i = 0; i < objectCount; i++) {
-            objectArray[i] = buffer.getFloat();
+            if (vectorType == VectorDimensionType.FLOAT16) {
+                // Server sends 2-byte float16, convert to 4-byte float for user
+                objectArray[i] = float16ToFloat(buffer.getShort());
+            } else {
+                objectArray[i] = buffer.getFloat(); // Read 4 bytes for FLOAT32
+            }
         }
 
         return new Vector(objectCount, vectorType, objectArray);
@@ -97,11 +93,12 @@ class VectorUtils {
 
         Object[] data = vector.getData();
         switch (vector.getVectorDimensionType()) {
-            // case FLOAT16:
-            //     for (Object value : data) {
-            //         buffer.putShort((short) ((Number) value).intValue());
-            //     }
-            //     break;
+            case FLOAT16:
+                for (Object value : data) {
+                    value = floatToFloat16((Float) value);
+                    buffer.putShort((short) ((Number) value).intValue());
+                }
+                break;
             case FLOAT32:
             default:
                 for (Object value : data) {
@@ -114,7 +111,7 @@ class VectorUtils {
     }
 
     static int getDefaultPrecision() {
-        return BYTES_PER_FLOAT; // FLOAT32
+        return BYTES_PER_FLOAT32; // FLOAT32
     }
 
     static int getHeaderLength() {
@@ -127,10 +124,10 @@ class VectorUtils {
      */
     static VectorDimensionType getVectorDimensionType(int scaleByte) {
         switch (scaleByte) {
-            case 0:
+            case SCALE_BYTE_FLOAT32:
                 return VectorDimensionType.FLOAT32;
-            // case 1:
-            // return VectorDimensionType.FLOAT16;
+            case SCALE_BYTE_FLOAT16:
+                return VectorDimensionType.FLOAT16;
             default:
                 return VectorDimensionType.FLOAT32;
         }
@@ -150,12 +147,12 @@ class VectorUtils {
      */
     static int getBytesPerDimensionFromScale(int scaleByte) {
         switch (scaleByte) {
-            case 0:
-                return BYTES_PER_FLOAT;
-            // case 1:
-            // return BYTES_PER_SHORT;
+            case SCALE_BYTE_FLOAT32:
+                return BYTES_PER_FLOAT32;
+            case SCALE_BYTE_FLOAT16:
+                return BYTES_PER_FLOAT16;
             default:
-                return BYTES_PER_FLOAT;
+                return BYTES_PER_FLOAT32;
         }
     }
 
@@ -165,11 +162,11 @@ class VectorUtils {
      */
     static int getBytesPerDimensionFromScale(VectorDimensionType vectorType) {
         switch (vectorType) {
-            // case FLOAT16:
-            // return BYTES_PER_SHORT;
+            case FLOAT16:
+                return BYTES_PER_FLOAT16;
             case FLOAT32:
             default:
-                return BYTES_PER_FLOAT;
+                return BYTES_PER_FLOAT32;
         }
     }
 
@@ -179,8 +176,8 @@ class VectorUtils {
      */
     static byte getScaleByte(VectorDimensionType vectorType) {
         switch (vectorType) {
-            // case FLOAT16:
-            // return SCALE_BYTE_FLOAT16;
+            case FLOAT16:
+                return SCALE_BYTE_FLOAT16;
             case FLOAT32:
             default:
                 return SCALE_BYTE_FLOAT32;
@@ -193,9 +190,9 @@ class VectorUtils {
      */
     static byte getScaleByte(int scale) {
         switch (scale) {
-            // case BYTES_PER_SHORT:
-            // return SCALE_BYTE_FLOAT16;
-            case BYTES_PER_FLOAT:
+            case BYTES_PER_FLOAT16:
+                return SCALE_BYTE_FLOAT16;
+            case BYTES_PER_FLOAT32:
             default:
                 return SCALE_BYTE_FLOAT32;
         }
@@ -208,12 +205,12 @@ class VectorUtils {
     static int getVectorLength(Vector vector) {
         int bytesPerDimension;
         switch (vector.getVectorDimensionType()) {
-            // case FLOAT16:
-            // bytesPerDimension = BYTES_PER_SHORT;
-            // break;
+            case FLOAT16:
+                bytesPerDimension = BYTES_PER_FLOAT16;
+                break;
             case FLOAT32:
             default:
-                bytesPerDimension = BYTES_PER_FLOAT;
+                bytesPerDimension = BYTES_PER_FLOAT32;
                 break;
         }
         return getHeaderLength() + bytesPerDimension * vector.getDimensionCount(); // 8-byte header + dimension payload
@@ -229,23 +226,48 @@ class VectorUtils {
 
     /**
      * Returns the SQL type definition string for a vector parameter.
-     * 
-     * @param vector      The vector instance (may be null for output-only
-     *                    parameters)
-     * @param scale       Number of bytes per dimension (e.g., 4 for FLOAT32, 2 for
-     *                    FLOAT16)
-     * @param isOutput    True if the parameter is an output parameter
-     * @param outScale    Output parameter's bytes per dimension (if applicable)
-     * @param valueLength The value length for output parameters
-     * @return SQL type definition string, e.g., VECTOR(128)
+     *
+     * Resolves the bytes-per-dimension scale internally:
+     * - For non-null vectors the scale is derived from the vector's dimension type
+     *   (e.g. FLOAT32 = 4, FLOAT16 = 2).
+     * - For null or output parameters, falls back to outScale (if output)
+     *   or the default precision.
+     *
+     * @param vector             The vector instance (may be null for null / output parameters)
+     * @param isOutput           Whether this is a registered output parameter
+     * @param outScale           The registered output scale (used only when isOutput is true)
+     * @param valueLength        The dimension count for null vector parameters
+     * @param negotiatedVersion  The negotiated vector version (1 for v1, 2 for v2)
+     * @return SQL type definition string, e.g., VECTOR(3, FLOAT16) or VECTOR(3)
      */
-    static String getTypeDefinition(Vector vector, int scale, boolean isOutput, int outScale, int valueLength) {
-        int precision = 0;
-        if (isOutput && scale < outScale) {
-            precision = valueLength;
-        } else if (vector != null) {
-            precision = vector.getDimensionCount();
+    static String getTypeDefinition(Vector vector, boolean isOutput, int outScale,
+            int valueLength, byte negotiatedVersion) {
+        int scale = (vector != null)
+                ? getBytesPerDimensionFromScale(vector.getVectorDimensionType())
+                : (isOutput ? outScale : getDefaultPrecision());
+
+        int precision = (vector != null) ? vector.getDimensionCount() : valueLength;
+
+        // Version 0: Server does not support vector at all
+        if (negotiatedVersion <= 0) {
+            throw vectorException("R_vectorNotSupported");
         }
+
+        // Version 1: Only FLOAT32 is supported; reject FLOAT16
+        if (negotiatedVersion == 1 && scale == BYTES_PER_FLOAT16) {
+            throw vectorException("R_float16VectorNotSupported");
+        }
+
+        // V2+: Explicitly specify type for both FLOAT32 and FLOAT16
+        //      VECTOR(dimensionCount, FLOAT32) or VECTOR(dimensionCount, FLOAT16)
+        if (negotiatedVersion >= 2) {
+            if (scale == BYTES_PER_FLOAT16) {
+                return "VECTOR(" + precision + ", FLOAT16)";
+            } else {
+                return "VECTOR(" + precision + ", FLOAT32)";
+            }
+        }
+        // V1: FLOAT32 only - use VECTOR(dimensionCount)
         return "VECTOR(" + precision + ")";
     }
