+++ src/main/java/microsoft/sql/Vector.java
@@ -10,10 +10,26 @@ import java.util.MissingResourceException;
 import java.util.ResourceBundle;
 import java.util.Arrays;
 
+/**
+ * Represents a SQL Server {@code VECTOR} data type value.
+ *
+ * A vector stores a fixed-length array of floating-point values, with support for
+ * {@link VectorDimensionType#FLOAT32 FLOAT32} (32-bit) and
+ * {@link VectorDimensionType#FLOAT16 FLOAT16} (16-bit half-precision) element types.
+ *
+ * Example usage:
+ * {@code
+ * Float[] data = new Float[] { 0.1f, 0.2f, 0.3f };
+ * Vector vector = new Vector(3, VectorDimensionType.FLOAT32, data);
+ * pstmt.setObject(1, vector, microsoft.sql.Types.VECTOR);
+ * }
+ *
+ * @since 13.3.2
+ */
 public final class Vector implements java.io.Serializable {
 
     public enum VectorDimensionType {
-        // FLOAT16, 16-bit (half precision) float
+        FLOAT16, // 16-bit (half precision) float
         FLOAT32 // 32-bit (single precision) float
     }
 
@@ -27,7 +43,7 @@ public final class Vector implements java.io.Serializable {
      * 
      * @param dimensionCount The number of dimensions in the vector.
      * @param vectorType     The type of the vector.
-     * @param data           The object array representing the vector data.
+     * @param data           The vector data. Must be {@code Float[]} or {@code null}.
      */
     public Vector(int dimensionCount, VectorDimensionType vectorType, Object[] data) {
         validateVectorParameters(dimensionCount, vectorType, data);
@@ -41,8 +57,8 @@ public final class Vector implements java.io.Serializable {
      * Constructor for Vector with precision and scale value.
      * 
      * @param precision The number of dimensions in the vector.
-     * @param scale     The scale value of the vector (4 for FLOAT32).
-     * @param data      The object array representing the vector data.
+     * @param scale     The scale value of the vector (4 for FLOAT32, 2 for FLOAT16).
+     * @param data      The vector data. Must be {@code Float[]} or {@code null}.
      */
     public Vector(int precision, int scale, Object[] data) {
         this(precision, getVectorDimensionTypeFromScaleValue(scale), data);
@@ -51,7 +67,7 @@ public final class Vector implements java.io.Serializable {
     /**
      * Returns the data of the vector.
      * 
-     * @return The object array representing the vector data.
+     * @return The vector data as {@code Float[]}, or {@code null}.
      */
     public Object[] getData() {
         return data;
@@ -82,8 +98,7 @@ public final class Vector implements java.io.Serializable {
      */
     @Override
     public String toString() {
-        return "VECTOR(" + vectorType + ", " + dimensionCount + ") : " +
-                (data != null ? Arrays.toString(data) : "null");
+        return "VECTOR(" + vectorType + ", " + dimensionCount + ") : " + Arrays.toString(data);
     }
 
     private static void validateVectorParameters(int dimensionCount, VectorDimensionType vectorType, Object[] data) {
@@ -121,8 +136,8 @@ public final class Vector implements java.io.Serializable {
         switch (scale) {
             case 4:
                 return VectorDimensionType.FLOAT32;
-            // case 2:
-            // return VectorDimensionType.FLOAT16;
+            case 2:
+                return VectorDimensionType.FLOAT16;
             default:
                 return VectorDimensionType.FLOAT32;
         }
